+++ schema-engine/connectors/sql-schema-connector/Cargo.toml
@@ -22,7 +22,9 @@ indexmap.workspace = true
 
 prisma-value = { path = "../../../libs/prisma-value" }
 schema-connector = { path = "../schema-connector" }
-sql-schema-describer = { path = "../../sql-schema-describer" }
+sql-schema-describer = { path = "../../sql-schema-describer", features = [
+    "all-native",
+] }
 datamodel-renderer = { path = "../../datamodel-renderer" }
 sql-ddl = { path = "../../../libs/sql-ddl" }
 user-facing-errors = { path = "../../../libs/user-facing-errors", features = [
